experiments/replication/pyproject.toml
[project]
name = "identity-replication"
version = "0.1.0"
description = "Clone identity test for persona-level replication (Appendix C)"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.50.0",
    "pydantic>=2.0.0",
    "typer>=0.12.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.0.0",
    "rich>=13.0.0",
    "scipy>=1.10.0",
    "pyyaml>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/clone_test"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
